Smaller, smarter, better method for creating case overlays.
Was trying to figure out out to apply it to other views (still not done
yet), and realized that the code was going to grow a LOT and be hard to
do.  This way the code can be easily added and with out a lot more
controls (1 per view) and only the type logo needs to be aligned per
view.  I think.