Added all of jingais fixes as needed.
Everything added to official was added except I kept my modified music
overlay (but added video on everything but homescreen, and added the
album year to the info displayed).  Didn't really need to add custom
weather as it already is (path option maybe later).
Added a default Apple trailers menu item if script is already installed
on the main menu defaults, but it is kind of silly as the user can add
whatever menu objects they want (or delete them).